[FIXED] pack rrr和b4r时pic存的原始大小
Former-commit-id: c4fff323eb92eef5000bf28a21c2f8a28fa76b2f

+++ epbin/src/BinaryB4R.h
@@ -21,7 +21,7 @@ public:
 private:
 	struct Picture
 	{
-		int bmp_w, bmp_h;
+		int16_t bmp_w, bmp_h;
 		uint8_t* bmp_pixels;
 
 		int16_t w, h;
